# apps/cloud-gateway/openapi.yaml
openapi: 3.1.0
info:
  title: RouteBox Cloud Gateway API
  version: 1.0.0
  description: |
    Multi-provider LLM routing gateway with OpenAI-compatible API.
    Routes requests to OpenAI, Anthropic, Google, DeepSeek, MiniMax, and Kimi
    with automatic retry, fallback, and circuit breaker for high availability.
  contact:
    name: RouteBox
    url: https://routebox.dev
servers:
  - url: https://cloud.routebox.dev
    description: Production
  - url: http://localhost:3001
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /account/me:
    get:
      tags: [Account]
      summary: Get full user profile with balance
      operationId: getAccountMe
      responses:
        "200":
          description: Full user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/AuthError"
        "404":
          description: User not found
  /account/balance:
    get:
      tags: [Account]
      summary: Get current credit balance
      operationId: getBalance
      responses:
        "200":
          description: Current balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance_cents:
                    type: integer
                    description: Credit balance in cents
        "401":
          $ref: "#/components/responses/AuthError"
  /account/transactions:
    get:
      tags: [Account]
      summary: Get transaction history
      operationId: getTransactions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Transaction"
        "401":
          $ref: "#/components/responses/AuthError"
  /account/referral:
    get:
      tags: [Account]
      summary: Get or create referral code
      operationId: getReferral
      responses:
        "200":
          description: Referral code and stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    description: 6-character referral code
                  uses:
                    type: integer
                  totalRewardCents:
                    type: integer
        "401":
          $ref: "#/components/responses/AuthError"
...
  /v1/models:
    get:
      tags: [AI]
      summary: List available models
      operationId: listModels
      responses:
        "200":
          description: Available models
          content:
            application/json:
              schema:
                type: object
                properties:
                  object:
                    type: string
                    enum: [list]
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        object:
                          type: string
                          enum: [model]
                        created:
                          type: integer
                        owned_by:
                          type: string
        "401":
          $ref: "#/components/responses/AuthError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
  responses:
    AuthError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [message, type]
          properties:
            message:
              type: string
            type:
              type: string
              enum:
                - auth_error
                - validation_error
                - conflict_error
                - billing_error
                - rate_limit_error
                - invalid_request_error
                - upstream_error
                - server_error
                - not_found
            code:
              type: string
            upstream:
              type: string
              description: Truncated upstream error body (for upstream_error type)
            retry_after_seconds:
              type: integer
              description: Seconds to wait before retrying (for rate_limit_error)
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
          nullable: true
        plan:
          type: string
          enum: [free, pro, business]
        balanceCents:
          type: integer
        totalDepositedCents:
          type: integer
        totalUsedCents:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [credit, debit, refund]
        amountCents:
          type: integer
        description:
          type: string
        createdAt:
          type: string
          format: date-time
...